Correct types in MFA claim
